codex-suite/requirements.txt
# Codex Dominion Suite - Core Dependencies

# Web Framework & API
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# Configuration & Environment
python-dotenv>=1.0.0

# Caching & Performance
redis>=5.0.0

# HTTP Client & Requests
requests>=2.31.0

# Vector Database & AI Search
faiss-cpu>=1.7.4

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.17.0

# AI and Machine Learning
openai>=1.3.0
anthropic>=0.7.0
transformers>=4.35.0
torch>=2.1.0

# Additional HTTP and API
aiohttp>=3.9.0
httpx>=0.25.0

# Database and Storage
json5>=0.9.0

# Utility Libraries
python-dateutil>=2.8.0
pytz>=2023.3
pathlib2>=2.3.0
typing-extensions>=4.8.0

# Security
cryptography>=41.0.0
pyjwt>=2.8.0
bcrypt>=4.1.0

# Development Tools
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Optional Enhanced Features
# Uncomment if using advanced AI models
# openai-gym>=0.26.0
# langchain>=0.0.350
# chromadb>=0.4.0

# Optional Video Processing (for Video Studio integration)
# opencv-python>=4.8.0
# pillow>=10.0.0
# moviepy>=1.0.0

# Optional Audio Processing
# pydub>=0.25.0
# speech-recognition>=3.10.0
# gtts>=2.4.0

# Optional Advanced Analytics
# scikit-learn>=1.3.0
# tensorflow>=2.14.0
# keras>=2.14.0

# Performance and Monitoring
psutil>=5.9.0
memory-profiler>=0.61.0
py-spy>=0.3.0

# Logging and Debugging
loguru>=0.7.0
rich>=13.6.0
typer>=0.9.0